requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mido==1.3.2
obs-websocket-py==1.0
packaging==23.2
python-rtmidi==1.5.8
websocket-client==1.8.0
websockets==13.0.1
Werkzeug==3.0.4
